Add flower stock-based product availability with multi-category products
- Products are many-to-many categories (category_product pivot); category
  string column removed from schema and migration data
- New flower_product pivot links products to flower stock; a product is
  unavailable when deactivated or any linked flower has 0 stock
- products.is_active admin toggle; availability badges and unavailable
  overlays across storefront, admin, cart guard, and customize page
- Regenerate database.sql dump

+++ database/migrations/2026_01_01_000007_create_order_items_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             Schema::create('order_items', function (Blueprint $table) {
                 $table->id();
                 $table->unsignedBigInteger('order_id');
-                $table->unsignedBigInteger('product_id')->nullable();
+                $table->integer('product_id')->nullable();
                 $table->string('product_name');
                 $table->decimal('price', 10, 2);
                 $table->unsignedInteger('quantity');
